+++ NEWS.md
@@ -163,6 +163,31 @@ Outstanding ones only.
             p C.ancestors #=> [C, M1, M2, Object, Kernel, BasicObject]
             ```
 
+    * New method
+
+      * Module#descendants, which returns a list of modules including
+        specified module or a list of subclasses inheriting from
+        specified class (including specified module/class itself).
+        [[Feature #14394]]
+
+        ```ruby
+        module A; end
+        module B; end
+        module C; end
+        B.include A
+        C.include B
+        p A.descendants    #=> [A, C, B]
+        p B.descendants    #=> [B, C]
+        p C.descendants    #=> [C]
+
+        class A; end
+        class B < A; end
+        class C < B; end
+        p A.descendants    #=> [A, B, C]
+        p B.descendants    #=> [B, C]
+        p C.descendants    #=> [C]
+        ```
+
 * Range
 
     * All Range objects are frozen. [Feature #15504]
@@ -411,7 +436,7 @@ Excluding feature bug fixes.
 
 * RBS is a new language for type definition of Ruby programs.
   It allows writing types of classes and modules with advanced
-  types including union types, overloading, generics, and 
+  types including union types, overloading, generics, and
   _interface types_ for duck typing.
 
 * Ruby ships with type definitions for core/stdlib classes.
@@ -471,3 +496,4 @@ Excluding feature bug fixes.
 [Feature #17122]: https://bugs.ruby-lang.org/issues/17122
 [Feature #17134]: https://bugs.ruby-lang.org/issues/17134
 [GH-2991]:        https://github.com/ruby/ruby/pull/2991
+[Feature #14394]: https://bugs.ruby-lang.org/issues/14394